added urls to each news mode
